requirements.txt
openai>=2.0.0
python-dotenv>=1.0.1
streamlit>=1.44.0
